requirements.txt
fastapi
uvicorn
pandas
beautifulsoup4
requests
pykakasi
keibascraper
python-dotenv
SQLAlchemy
